добавил таймер

+++ src/Game/GameObject/Eagle.cpp
@@ -17,6 +17,6 @@ void Eagle::render() const
 	m_Sprites[static_cast<size_t>(m_currentState)]->render(m_position, m_size, m_rotation, m_layer);
 }
 
-void Eagle::update(const uint64_t delta)
+void Eagle::update(const double delta)
 {
 }
