+++ src/Game/GameObject/Eagle.h
@@ -23,7 +23,7 @@ public:
 	Eagle(const glm::vec2& position, const glm::vec2& size, const float rotation, const float layer);
 
 	void render() const override;
-	void update(const uint64_t) override;
+	void update(const double) override;
 
 private:
 	std::array<std::shared_ptr<RenderEngine::Sprite>, 2> m_Sprites;
