+++ src/Game/GameObject/Water.cpp
@@ -23,7 +23,7 @@ void Water::render() const
 	renderWater(EWaterLocation::BottomRight);
 }
 
-void Water::update(const uint64_t delta)
+void Water::update(const double delta)
 {
 	m_spriteAnimator.update(delta);
 }
