+++ src/Game/Level.cpp
@@ -154,7 +154,7 @@ void Level::render() const
 	}
 }
 
-void Level::update(const uint64_t delta)
+void Level::update(const double delta)
 {
 	for (const auto currentObject : m_mapObject)
 	{
