+++ src/Renderer/Sprite.cpp
@@ -101,7 +101,7 @@ namespace RenderEngine
 
 	}
 
-	uint64_t Sprite::getFrameDuration(const size_t frameId) const
+	double Sprite::getFrameDuration(const size_t frameId) const
 	{
 		return m_framesDescriptions[frameId].duration;
 	}
